refactor(task-2): change solution structure
- use `queryKeys` provided with the project
- extract Budget Table to separated component `<BudgetTableWidget />`
- create function for getting amount status
- provide prop for `<Error />` component
