+++ client/src/hooks/useBudgets.js
@@ -1,6 +1,7 @@
 import { BudgetService } from 'api';
+import { BUDGET_QUERY } from 'queryKeys';
 import { useQuery } from 'react-query';
 
 const getBudgets = BudgetService.findAll;
 
-export const useBudgets = () => useQuery('budgets', getBudgets);
+export const useBudgets = () => useQuery(BUDGET_QUERY, getBudgets);
