+++ client/src/hooks/useDeleteBudget.js
@@ -1,4 +1,5 @@
 import { BudgetService } from 'api';
+import { BUDGET_QUERY } from 'queryKeys';
 import { useMutation, useQueryClient } from 'react-query';
 
 const deleteBudget = (values) => BudgetService.remove({ ids: values });
@@ -7,8 +8,8 @@ export const useDeleteBudget = () => {
   const queryClient = useQueryClient();
 
   return useMutation(deleteBudget, {
-    onSuccess: () => {
-      queryClient.invalidateQueries('budgets');
+    onSuccess: async () => {
+      await queryClient.invalidateQueries(BUDGET_QUERY);
     },
   });
 };
